Files and folder reorganization
