+++ src/Plot/Plot.jl
@@ -1,0 +1,10 @@
+module Plot
+
+include("plot.jl")
+
+export
+    wplotdots,
+    wplotim
+
+
+end
