+++ src/WT/WT.jl
@@ -1,0 +1,14 @@
+module WT
+
+include("wt.jl")
+
+export
+    DiscreteWavelet,
+    FilterWavelet,
+    LSWavelet,
+    OrthoFilter,
+    GLS,
+    wavelet
+
+
+end
